Merge was perfomed with 3 merge commands:
svn merge -r 4225:4226 https://svn.janelia.org/eddylab/eddys/src/infernal/branches/1.1
svn merge -r 4230:4231 https://svn.janelia.org/eddylab/eddys/src/infernal/branches/1.1
svn merge -r 4231:4232 https://svn.janelia.org/eddylab/eddys/src/infernal/branches/1.1


git-svn-id: https://svn.janelia.org/eddylab/eddys/src/infernal/trunk@4233 db3e70e4-12e1-0310-b23f-81a6e3d09ba7

+++ src/cm_pipeline.c
@@ -987,7 +987,7 @@ cm_pli_NewModel(CM_PIPELINE *pli, int modmode, CM_t *cm, int cm_clen, int cm_W, 
       }
     }
     else { /* ! do_hmmonly_cur */
-      if((status = UpdateExpsForDBSize(cm, pli->errbuf, (long) pli->Z)) != eslOK) return status;
+      if((status = UpdateExpsForDBSize(cm, pli->errbuf, pli->Z)) != eslOK) return status;
       if(pli->by_E) { 
 	if((status = E2ScoreGivenExpInfo(cm->expA[pli->final_cm_exp_mode], pli->errbuf, pli->E, &T)) != eslOK) ESL_FAIL(status, pli->errbuf, "problem determining min score for E-value %6g for model %s\n", pli->E, cm->name);
 	pli->T = (double) T;
